[CHORE] extension정리
