+++ Gajeongtongsin/Gajeongtongsin/Global/Extension/Date+Extension.swift
@@ -1,0 +1,17 @@
+//
+//  Date+Extension.swift
+//  Gajeongtongsin
+//
+//  Created by uiskim on 2022/07/29.
+//
+
+import Foundation
+
+extension Date {
+    func toString() -> String {
+        let dateFormatter = DateFormatter()
+        dateFormatter.dateFormat = "MM월dd일"
+        dateFormatter.timeZone = TimeZone(identifier: "ko_KR")
+        return dateFormatter.string(from: self)
+    }
+}
